Refactor timeline events to separate file
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/acrobid/negotiation-dates-bolt-new?shareId=XXXX-XXXX-XXXX-XXXX).
