+++ src/data/timelineEvents.ts
@@ -1,0 +1,94 @@
+interface TimelineEvent {
+  dateRange: {
+    start: Date;
+    end: Date;
+  };
+  description?: string;
+  markdownFile?: string;
+  canceled?: boolean;
+}
+
+const timelineEvents: TimelineEvent[] = [
+  {
+    dateRange: {
+      start: new Date(2024, 9, 21), // October 21
+      end: new Date(2024, 9, 24), // October 24
+    },
+    markdownFile: "update1.md",
+  },
+  {
+    dateRange: {
+      start: new Date(2024, 10, 12), // November 12
+      end: new Date(2024, 10, 14, 23), // November 14
+    },
+    markdownFile: "update2.md",
+  },
+  {
+    dateRange: {
+      start: new Date(2024, 10, 19), // November 19
+      end: new Date(2024, 10, 22, 23), // November 22
+    },
+    markdownFile: "update3.md",
+  },
+  {
+    dateRange: {
+      start: new Date(2024, 11, 10), // December 10
+      end: new Date(2024, 11, 12, 23), // December 12
+    },
+    markdownFile: "update4.md",
+  },
+  {
+    dateRange: {
+      start: new Date(2025, 0, 14), // January 14
+      end: new Date(2025, 0, 16, 23), // January 16
+    },
+    description: "Canceled by the national mediation board.",
+    canceled: true,
+  },
+  {
+    dateRange: {
+      start: new Date(2025, 1, 4), // February 4
+      end: new Date(2025, 1, 7, 23), // February 7
+    },
+    description:
+      "Preempted for the February 11th status update meeting with the NMB.",
+    canceled: true,
+  },
+  {
+    dateRange: {
+      start: new Date(2025, 1, 11), // February 11
+      end: new Date(2025, 1, 11), // February 11
+    },
+    description: "Status update meeting with new NMB mediator.",
+  },
+  {
+    dateRange: {
+      start: new Date(2025, 3, 15), // April 15
+      end: new Date(2025, 3, 17), // April 17
+    },
+    description: "Negotiation session.",
+  },
+  {
+    dateRange: {
+      start: new Date(2025, 3, 29), // April 29
+      end: new Date(2025, 4, 1), // May 1
+    },
+    description: "Negotiation session.",
+  },
+  {
+    dateRange: {
+      start: new Date(2025, 4, 28), // May 28
+      end: new Date(2025, 4, 30), // May 30
+    },
+    description: "Negotiation session.",
+  },
+  {
+    dateRange: {
+      start: new Date(2025, 5, 24), // June 24
+      end: new Date(2025, 5, 26), // June 26
+    },
+    description: "Negotiation session.",
+  },
+];
+
+export { timelineEvents, TimelineEvent };
